+++ miden-lib/src/tests/test_faucet.rs
@@ -1,12 +1,18 @@
 use miden_objects::assets::FungibleAsset;
 use mock::{
     constants::{
-        non_fungible_asset, non_fungible_asset_2, ACCOUNT_ID_FUNGIBLE_FAUCET_ON_CHAIN,
-        ACCOUNT_ID_FUNGIBLE_FAUCET_ON_CHAIN_1, ACCOUNT_ID_NON_FUNGIBLE_FAUCET_ON_CHAIN,
-        ACCOUNT_ID_NON_FUNGIBLE_FAUCET_ON_CHAIN_1, CONSUMED_ASSET_1_AMOUNT, FUNGIBLE_ASSET_AMOUNT,
+        non_fungible_asset, non_fungible_asset_2, CONSUMED_ASSET_1_AMOUNT, FUNGIBLE_ASSET_AMOUNT,
         FUNGIBLE_FAUCET_INITIAL_BALANCE,
     },
-    mock::{account::MockAccountType, notes::AssetPreservationStatus, transaction::mock_inputs},
+    mock::{
+        account::{
+            MockAccountType, ACCOUNT_ID_FUNGIBLE_FAUCET_ON_CHAIN,
+            ACCOUNT_ID_FUNGIBLE_FAUCET_ON_CHAIN_1, ACCOUNT_ID_NON_FUNGIBLE_FAUCET_ON_CHAIN,
+            ACCOUNT_ID_NON_FUNGIBLE_FAUCET_ON_CHAIN_1,
+        },
+        notes::AssetPreservationStatus,
+        transaction::mock_inputs,
+    },
     prepare_transaction,
     procedures::prepare_word,
     run_tx,
