+++ mock/src/builders/account_id.rs
@@ -7,7 +7,7 @@ use rand::Rng;
 
 use crate::{
     builders::{str_to_account_code, AccountBuilderError},
-    constants::DEFAULT_ACCOUNT_CODE,
+    mock::account::DEFAULT_ACCOUNT_CODE,
 };
 
 /// Builder for an `AccountId`, the builder can be configured and used multiple times.
