+++ mock/src/constants.rs
@@ -1,55 +1,18 @@
 use miden_objects::{
-    accounts::{get_account_seed_single, AccountId, AccountType, SlotItem, StorageSlotType},
+    accounts::AccountId,
     assets::{Asset, NonFungibleAsset, NonFungibleAssetDetails},
-    Felt, FieldElement, Word, ZERO,
+    Felt,
 };
 
 pub use super::mock::account::{
     ACCOUNT_PROCEDURE_INCR_NONCE_PROC_IDX, ACCOUNT_PROCEDURE_SET_CODE_PROC_IDX,
     ACCOUNT_PROCEDURE_SET_ITEM_PROC_IDX,
 };
-use super::{
-    mock::account::{mock_account, mock_fungible_faucet, mock_non_fungible_faucet},
-    TransactionKernel,
-};
-use crate::mock::account::mock_account_code;
 
-pub const ACCOUNT_ID_REGULAR_ACCOUNT_UPDATABLE_CODE_ON_CHAIN: u64 = 3238098370154045919;
-
-pub const ACCOUNT_ID_SENDER: u64 = 0b0110111011u64 << 54;
-pub const ACCOUNT_ID_FUNGIBLE_FAUCET_ON_CHAIN: u64 = 0b1010111100 << 54;
-pub const ACCOUNT_ID_NON_FUNGIBLE_FAUCET_ON_CHAIN: u64 = 0b1110011100 << 54;
-pub const ACCOUNT_ID_NON_FUNGIBLE_FAUCET_ON_CHAIN_1: u64 = 0b1110011101 << 54;
 pub const FUNGIBLE_ASSET_AMOUNT: u64 = 100;
 pub const FUNGIBLE_FAUCET_INITIAL_BALANCE: u64 = 50000;
 
-pub const ACCOUNT_ID_FUNGIBLE_FAUCET_ON_CHAIN_1: u64 =
-    0b1010010001111111010110100011011110101011010001101111110110111100u64;
-pub const ACCOUNT_ID_FUNGIBLE_FAUCET_ON_CHAIN_2: u64 =
-    0b1010000101101010101101000110111101010110100011011110100011011101u64;
-pub const ACCOUNT_ID_FUNGIBLE_FAUCET_ON_CHAIN_3: u64 =
-    0b1010011001011010101101000110111101010110100011011101000110111100u64;
-
 pub const MIN_PROOF_SECURITY_LEVEL: u32 = 96;
-
-// Default account code
-pub const DEFAULT_ACCOUNT_CODE: &str = "
-    use.miden::contracts::wallets::basic->basic_wallet
-    use.miden::contracts::auth::basic->basic_eoa
-
-    export.basic_wallet::receive_asset
-    export.basic_wallet::send_asset
-    export.basic_eoa::auth_tx_rpo_falcon512
-";
-
-// Default authentication code
-pub const DEFAULT_AUTH_SCRIPT: &str = "
-    use.miden::contracts::auth::basic->auth_tx
-
-    begin
-        call.auth_tx::auth_tx_rpo_falcon512
-    end
-";
 
 pub const CONSUMED_ASSET_1_AMOUNT: u64 = 100;
 pub const CONSUMED_ASSET_2_AMOUNT: u64 = 200;
@@ -58,21 +21,6 @@ pub const CONSUMED_ASSET_4_AMOUNT: u64 = 100;
 
 pub const NON_FUNGIBLE_ASSET_DATA: [u8; 4] = [1, 2, 3, 4];
 pub const NON_FUNGIBLE_ASSET_DATA_2: [u8; 4] = [5, 6, 7, 8];
-
-pub const NONCE: Felt = Felt::ZERO;
-
-pub const STORAGE_INDEX_0: u8 = 20;
-pub const STORAGE_VALUE_0: Word = [Felt::new(1), Felt::new(2), Felt::new(3), Felt::new(4)];
-pub const STORAGE_INDEX_1: u8 = 30;
-pub const STORAGE_VALUE_1: Word = [Felt::new(5), Felt::new(6), Felt::new(7), Felt::new(8)];
-
-pub fn storage_item_0() -> SlotItem {
-    (STORAGE_INDEX_0, (StorageSlotType::Value { value_arity: 0 }, STORAGE_VALUE_0))
-}
-
-pub fn storage_item_1() -> SlotItem {
-    (STORAGE_INDEX_1, (StorageSlotType::Value { value_arity: 0 }, STORAGE_VALUE_1))
-}
 
 pub const CHILD_ROOT_PARENT_LEAF_INDEX: u8 = 10;
 pub const CHILD_SMT_DEPTH: u8 = 64;
@@ -99,81 +47,4 @@ pub fn non_fungible_asset_2(account_id: u64) -> Asset {
     let non_fungible_asset_2: NonFungibleAsset =
         NonFungibleAsset::new(&non_fungible_asset_2_details).unwrap();
     Asset::NonFungible(non_fungible_asset_2)
-}
-
-// ACCOUNT SEED GENERATION
-// ================================================================================================
-
-pub enum AccountSeedType {
-    FungibleFaucetInvalidInitialBalance,
-    FungibleFaucetValidInitialBalance,
-    NonFungibleFaucetInvalidReservedSlot,
-    NonFungibleFaucetValidReservedSlot,
-    RegularAccountUpdatableCodeOnChain,
-}
-
-/// Returns the account id and seed for the specified account type.
-pub fn generate_account_seed(account_seed_type: AccountSeedType) -> (AccountId, Word) {
-    let assembler = TransactionKernel::assembler();
-    let init_seed: [u8; 32] = Default::default();
-
-    let (account, account_type) = match account_seed_type {
-        AccountSeedType::FungibleFaucetInvalidInitialBalance => (
-            mock_fungible_faucet(
-                ACCOUNT_ID_REGULAR_ACCOUNT_UPDATABLE_CODE_ON_CHAIN,
-                ZERO,
-                false,
-                &assembler,
-            ),
-            AccountType::FungibleFaucet,
-        ),
-        AccountSeedType::FungibleFaucetValidInitialBalance => (
-            mock_fungible_faucet(
-                ACCOUNT_ID_REGULAR_ACCOUNT_UPDATABLE_CODE_ON_CHAIN,
-                ZERO,
-                true,
-                &assembler,
-            ),
-            AccountType::FungibleFaucet,
-        ),
-        AccountSeedType::NonFungibleFaucetInvalidReservedSlot => (
-            mock_non_fungible_faucet(
-                ACCOUNT_ID_REGULAR_ACCOUNT_UPDATABLE_CODE_ON_CHAIN,
-                ZERO,
-                false,
-                &assembler,
-            ),
-            AccountType::NonFungibleFaucet,
-        ),
-        AccountSeedType::NonFungibleFaucetValidReservedSlot => (
-            mock_non_fungible_faucet(
-                ACCOUNT_ID_REGULAR_ACCOUNT_UPDATABLE_CODE_ON_CHAIN,
-                ZERO,
-                true,
-                &assembler,
-            ),
-            AccountType::NonFungibleFaucet,
-        ),
-        AccountSeedType::RegularAccountUpdatableCodeOnChain => (
-            mock_account(
-                ACCOUNT_ID_REGULAR_ACCOUNT_UPDATABLE_CODE_ON_CHAIN,
-                Felt::ONE,
-                mock_account_code(&assembler),
-            ),
-            AccountType::RegularAccountUpdatableCode,
-        ),
-    };
-
-    let seed = get_account_seed_single(
-        init_seed,
-        account_type,
-        true,
-        account.code().root(),
-        account.storage().root(),
-    )
-    .unwrap();
-
-    let account_id = AccountId::new(seed, account.code().root(), account.storage().root()).unwrap();
-
-    (account_id, seed)
 }
